Refactor layout management and controls
- Removed the old `useLayout.js` and replaced it with a new implementation that integrates with the Vue Flow library.
- Introduced `Controls.vue` for managing layout controls, including save/restore functionality and layout direction options.
- Added `Header.vue` for a consistent header layout with save/restore controls.
- Created `Icon.vue` for reusable SVG icons used in layout controls.
- Developed `LayoutControls.vue` to encapsulate layout-related buttons and functionality.
- Implemented `SaveRestoreControls.vue` for saving and restoring graph states with template support.
- Added drag-and-drop functionality in `useDnD.js` for various node types.
- Updated styles for better UI consistency and usability.
