+++ frontend/src/composables/useDnD.js
@@ -1,32 +1,32 @@
 import { v4 as uuidv4 } from 'uuid'
 import { useVueFlow } from '@vue-flow/core'
 import { ref, watch, nextTick } from 'vue'
-import NoteNode from './components/nodes/NoteNode.vue'
-import PythonRunner from './components/nodes/PythonRunner.vue'
-import WebGLNode from './components/nodes/WebGLNode.vue'
-import AgentNode from './components/nodes/AgentNode.vue'
-import ClaudeNode from './components/nodes/ClaudeNode.vue'
-import ResponseNode from './components/nodes/ResponseNode.vue'
-import GeminiNode from './components/nodes/GeminiNode.vue'
-import EmbeddingsNode from './components/nodes/EmbeddingsNode.vue'
-import WebSearchNode from './components/nodes/WebSearchNode.vue'
-import WebRetrievalNode from './components/nodes/WebRetrievalNode.vue'
-import TextNode from './components/nodes/TextNode.vue'
-import TextSplitterNode from './components/nodes/TextSplitterNode.vue'
-import OpenFileNode from './components/nodes/OpenFileNode.vue'
-import SaveTextNode from './components/nodes/SaveTextNode.vue'
-import DatadogNode from './components/nodes/DatadogNode.vue'
-import DatadogGraphNode from './components/nodes/DatadogGraphNode.vue'
-import TokenCounterNode from './components/nodes/TokenCounterNode.vue'
-import FlowControl from './components/FlowControl.vue'
-import RepoConcat from './components/nodes/RepoConcat.vue'
-import ComfyNode from './components/nodes/ComfyNode.vue'
-import MLXFlux from './components/nodes/MLXFlux.vue'
-import DocumentsIngest from './components/nodes/DocumentsIngestNode.vue'
-import DocumentsRetrieve from './components/nodes/DocumentsRetrieveNode.vue'
-import ttsNode from './components/nodes/ttsNode.vue'
-import MCPClientNode from './components/nodes/MCPClient.vue'
-import Mermaid from './components/nodes/Mermaid.vue'
+import NoteNode from '../components/nodes/NoteNode.vue'
+import PythonRunner from '../components/nodes/PythonRunner.vue'
+import WebGLNode from '../components/nodes/WebGLNode.vue'
+import AgentNode from '../components/nodes/AgentNode.vue'
+import ClaudeNode from '../components/nodes/ClaudeNode.vue'
+import ResponseNode from '../components/nodes/ResponseNode.vue'
+import GeminiNode from '../components/nodes/GeminiNode.vue'
+import EmbeddingsNode from '../components/nodes/EmbeddingsNode.vue'
+import WebSearchNode from '../components/nodes/WebSearchNode.vue'
+import WebRetrievalNode from '../components/nodes/WebRetrievalNode.vue'
+import TextNode from '../components/nodes/TextNode.vue'
+import TextSplitterNode from '../components/nodes/TextSplitterNode.vue'
+import OpenFileNode from '../components/nodes/OpenFileNode.vue'
+import SaveTextNode from '../components/nodes/SaveTextNode.vue'
+import DatadogNode from '../components/nodes/DatadogNode.vue'
+import DatadogGraphNode from '../components/nodes/DatadogGraphNode.vue'
+import TokenCounterNode from '../components/nodes/TokenCounterNode.vue'
+import FlowControl from '../components/FlowControl.vue'
+import RepoConcat from '../components/nodes/RepoConcat.vue'
+import ComfyNode from '../components/nodes/ComfyNode.vue'
+import MLXFlux from '../components/nodes/MLXFlux.vue'
+import DocumentsIngest from '../components/nodes/DocumentsIngestNode.vue'
+import DocumentsRetrieve from '../components/nodes/DocumentsRetrieveNode.vue'
+import ttsNode from '../components/nodes/ttsNode.vue'
+import MCPClientNode from '../components/nodes/MCPClient.vue'
+import Mermaid from '../components/nodes/Mermaid.vue'
 
 let id = 0
